backend/requirements.txt
alembic>=1.18.4
argon2-cffi>=25.1.0
asyncpg>=0.31.0
bcrypt>=5.0.0
fastapi>=0.135.1
httpx>=0.28.1
passlib>=1.7.4
pydantic[email]>=2.12.5
python-jose>=3.5.0
slowapi>=0.1.9
sqlalchemy>=2.0.48
uvicorn>=0.41.0
python-multipart
